add RecipientsViews component to ManageRecipients
